fix proposal actions ui issues (#162)
* chore: wip proposals

* chore: wip proposals

* chore: wip

* chore: wip

* chore: wip

* fix cancel proposal

* chore: wip

* chore: wip sign off proposal

* chore: wip sign off proposal

* chore: wip finalize votes

* feat: finalize votes

* feat: sign off proposal

* feat: sign off proposal

* feat: execute instruction

* feat: execute instruction

* chore: wip

* chore: wip

* chore: finalize vote

* fix: build errors

* chore: remove unused code

* chore: remove unused code

* chore: remove unused code

* fix: sign off proposal

* fix: execute instruction

* fix: execute instruction

* chore: wip

* chore: fix ux issues

* chore: wip ux issues

* chore: wip ux issues

* chore: ux issues

* chore: wip ux issues

* fix: cancel button enabled incorrectly

* chore: fix cancel button issue

* chore: fix cancel button

* fix: display Cancel button for proposal owner only

* chore: wip

* chore: wip

* chore: fix ui issues

* chore: fix ui issues

* chore: fix ui issues

* chore: fix ui issues

* chore: fix ui issues

* chore: fix ui issues

* chore: fix ui issues

* chore: fix ui issues on comment modal
