+++ .gitignore
@@ -18,6 +18,7 @@
 # misc
 .DS_Store
 *.pem
+.env
 
 # debug
 npm-debug.log*
