+++ components/CancelProposalModal.tsx
@@ -18,7 +18,7 @@ type CancelProposalModalProps = {
 const CancelProposalModal = ({ onClose, isOpen }: CancelProposalModalProps) => {
   const wallet = useWalletStore((s) => s.current)
   const connection = useWalletStore((s) => s.connection)
-  const fetchRealm = useWalletStore((s) => s.actions.fetchRealm)
+  const fetchProposal = useWalletStore((s) => s.actions.fetchProposal)
   const { realmInfo } = useRealm()
   const { proposal } = useProposal()
 
@@ -39,7 +39,7 @@ const CancelProposalModal = ({ onClose, isOpen }: CancelProposalModalProps) => {
 
         onClose()
 
-        await fetchRealm(realmInfo.programId, realmInfo.realmId)
+        await fetchProposal(proposal.pubkey)
       }
     } catch (error) {
       notify({
