+++ components/VotePanel.tsx
@@ -1,19 +1,18 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
+
 import { withFinalizeVote } from '@models/withFinalizeVote'
 import { TransactionInstruction } from '@solana/web3.js'
-import { useCallback, useEffect, useState } from 'react'
+import { useCallback, useState } from 'react'
 import { relinquishVote } from '../actions/relinquishVote'
 import { useHasVoteTimeExpired } from '../hooks/useHasVoteTimeExpired'
 import useRealm from '../hooks/useRealm'
-import { getSignatoryRecordAddress, ProposalState } from '../models/accounts'
+import { ProposalState } from '../models/accounts'
 import { RpcContext } from '../models/core/api'
 import { GoverningTokenType } from '../models/enums'
 
 import { Vote } from '../models/instructions'
 import useWalletStore from '../stores/useWalletStore'
 import Button from './Button'
-import CancelProposalModal from './CancelProposalModal'
-import FinalizeVotesModal from './FinalizeVotesModal'
-import SignOffProposalModal from './SignOffProposalModal'
 import VoteCommentModal from './VoteCommentModal'
 
 const VotePanel = () => {
@@ -24,7 +23,6 @@ const VotePanel = () => {
     proposal,
     voteRecordsByVoter,
     tokenType,
-    proposalOwner,
   } = useWalletStore((s) => s.selectedProposal)
   const { ownTokenRecord, ownCouncilTokenRecord, realm, realmInfo } = useRealm()
   const wallet = useWalletStore((s) => s.current)
@@ -32,37 +30,6 @@ const VotePanel = () => {
   const connected = useWalletStore((s) => s.connected)
   const fetchRealm = useWalletStore((s) => s.actions.fetchRealm)
   const hasVoteTimeExpired = useHasVoteTimeExpired(governance, proposal!)
-  const signatories = useWalletStore((s) => s.selectedProposal.signatories)
-
-  const [showSignOffModal, setShowSignOffModal] = useState(false)
-  const [signatoryRecord, setSignatoryRecord] = useState<any>(undefined)
-  const [showFinalizeVoteModal, setShowFinalizeVoteModal] = useState(false)
-  const [showCancelModal, setShowCancelModal] = useState(false)
-
-  const canFinalizeVote =
-    hasVoteTimeExpired === true &&
-    connected &&
-    proposal?.info.state === ProposalState.Voting
-
-  const walletPk = wallet?.publicKey
-
-  useEffect(() => {
-    const setup = async () => {
-      if (proposal && realmInfo && walletPk) {
-        const signatoryRecordPk = await getSignatoryRecordAddress(
-          realmInfo.programId,
-          proposal.pubkey,
-          walletPk
-        )
-
-        if (signatoryRecordPk && signatories) {
-          setSignatoryRecord(signatories[signatoryRecordPk.toBase58()])
-        }
-      }
-    }
-
-    setup()
-  }, [proposal, realmInfo, walletPk])
 
   const ownVoteRecord =
     wallet?.publicKey && voteRecordsByVoter[wallet.publicKey.toBase58()]
@@ -121,6 +88,7 @@ const VotePanel = () => {
       await relinquishVote(
         rpcContext,
         proposal!,
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         voterTokenRecord!.pubkey,
         ownVoteRecord!.pubkey,
         instructions
@@ -147,125 +115,57 @@ const VotePanel = () => {
     ? 'Withdraw your vote'
     : 'Release your tokens'
 
-  const canSignOff =
-    signatoryRecord &&
-    (proposal?.info.state === ProposalState.Draft ||
-      proposal?.info.state === ProposalState.SigningOff)
-
-  const canCancelProposal =
-    proposal &&
-    governance &&
-    proposalOwner &&
-    wallet?.publicKey &&
-    proposal.info.canWalletCancel(
-      governance.info,
-      proposalOwner.info,
-      wallet.publicKey
-    )
-
   return (
-    <div className="bg-bkg-2 p-4 md:p-6 rounded-lg space-y-6">
-      <h2 className="mb-4 text-center">{actionLabel}</h2>
-      <div
-        className={`${
-          isVoting && 'flex-col'
-        } flex justify-center items-center gap-5`}
-      >
-        {isVoteCast ? (
-          <Button
-            onClick={() => submitRelinquishVote()}
-            disabled={!isWithdrawEnabled}
-          >
-            {isVoting ? 'Withdraw' : 'Release Tokens'}
-          </Button>
-        ) : (
-          <>
-            {isVoting && (
-              <div className="border-b border-gray-600 flex gap-x-5 pb-6 w-full justify-center items-center">
-                <Button
-                  className="w-full"
-                  onClick={() => handleShowVoteModal(Vote.Yes)}
-                  disabled={!isVoteEnabled}
-                >
-                  Approve
-                </Button>
-                <Button
-                  className="w-full"
-                  onClick={() => handleShowVoteModal(Vote.No)}
-                  disabled={!isVoteEnabled}
-                >
-                  Deny
-                </Button>
-              </div>
+    <>
+      {ProposalState.Cancelled === proposal?.info.state ||
+      ProposalState.Succeeded === proposal?.info.state ||
+      ProposalState.Draft === proposal?.info.state ? null : (
+        <div className="bg-bkg-2 p-4 md:p-6 rounded-lg space-y-6">
+          <h2 className="mb-4 text-center">{actionLabel}</h2>
+
+          <div className="items-center justify-center flex w-full gap-5">
+            {isVoteCast ? (
+              <Button
+                onClick={() => submitRelinquishVote()}
+                disabled={!isWithdrawEnabled}
+              >
+                {isVoting ? 'Withdraw' : 'Release Tokens'}
+              </Button>
+            ) : (
+              <>
+                {isVoting && (
+                  <div className="w-full flex justify-between items-center gap-5">
+                    <Button
+                      className="w-1/2"
+                      onClick={() => handleShowVoteModal(Vote.Yes)}
+                      disabled={!isVoteEnabled}
+                    >
+                      Approve
+                    </Button>
+                    <Button
+                      className="w-1/2"
+                      onClick={() => handleShowVoteModal(Vote.No)}
+                      disabled={!isVoteEnabled}
+                    >
+                      Deny
+                    </Button>
+                  </div>
+                )}
+              </>
             )}
-          </>
-        )}
+          </div>
 
-        {canSignOff && (
-          <Button
-            className={isVoting ? 'w-1/2' : 'w-full'}
-            onClick={() => setShowSignOffModal(true)}
-            disabled={!connected || !canSignOff}
-          >
-            Sign Off
-          </Button>
-        )}
-
-        {canCancelProposal && (
-          <Button
-            className={isVoting ? 'w-1/2' : 'w-full'}
-            onClick={() => setShowCancelModal(true)}
-            disabled={!connected}
-          >
-            Cancel
-          </Button>
-        )}
-
-        {canFinalizeVote && (
-          <Button
-            className={isVoting ? 'w-full' : ''}
-            onClick={() => setShowFinalizeVoteModal(true)}
-            disabled={!connected || !canFinalizeVote}
-          >
-            Finalize
-          </Button>
-        )}
-      </div>
-
-      {showVoteModal ? (
-        <VoteCommentModal
-          isOpen={showVoteModal}
-          onClose={handleCloseShowVoteModal}
-          vote={vote!}
-          voterTokenRecord={voterTokenRecord!}
-        />
-      ) : null}
-
-      {showSignOffModal && (
-        <SignOffProposalModal
-          isOpen={showSignOffModal && canSignOff}
-          onClose={() => setShowSignOffModal(false)}
-          signatoryRecord={signatoryRecord}
-        />
+          {showVoteModal ? (
+            <VoteCommentModal
+              isOpen={showVoteModal}
+              onClose={handleCloseShowVoteModal}
+              vote={vote!}
+              voterTokenRecord={voterTokenRecord!}
+            />
+          ) : null}
+        </div>
       )}
-
-      {showFinalizeVoteModal && (
-        <FinalizeVotesModal
-          isOpen={showFinalizeVoteModal && canFinalizeVote}
-          onClose={() => setShowFinalizeVoteModal(false)}
-          proposal={proposal}
-          governance={governance}
-        />
-      )}
-
-      {showCancelModal && (
-        <CancelProposalModal
-          // @ts-ignore
-          isOpen={showCancelModal && canCancelProposal}
-          onClose={() => setShowCancelModal(false)}
-        />
-      )}
-    </div>
+    </>
   )
 }
